+++ changes.md
@@ -3,6 +3,11 @@
 ## DO NOT add new events at last. Recent changes should be at the top.
 
 
+### By <b>_Harshit Garg_</b> on <b>_October 27, 2020_</b>
+* Comment section for videos added.
+* Flag on videos and comments added.
+* Add videos to favourites.
+
 ### By <b>_Harshit Garg_</b> on <b>_October 26, 2020_</b>
 * Playing video through backend.
 * user dashboard with movies and series flow added.